requirements.txt
django
nltk
pdfminer.six
docx2txt
spacy
openpyxl
